Fix getTypes(from:pokemonDataStore:) bug
Fixed a bug where getTypes(from:pokemonDataStore:) would check for the
wrong thing in the data store. It would check for items instead
of types. Also fixed the issue where if a type with the given name was
found it would return nil. It now correctly returns the type that was
found.